refactor: split toon_parse.go and toon_tokenizer.go to stay under 500 LOC
Move header parsing logic to toon_header.go (250 LOC) and content
emission methods to toon_emit.go (262 LOC). Both source files now
comply with the 500 LOC convention.
